Refactor file structure for clarify
